# refactor: use b.Loop()