--- a/types/context_bench_test.go
+++ b/types/context_bench_test.go
@@ -13,8 +13,7 @@ func BenchmarkContext_KVStore(b *testing.B) {
 
 	ctx := testutil.DefaultContext(key, types.NewTransientStoreKey("transient_"+b.Name()))
 
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
 		_ = ctx.KVStore(key)
 	}
 }
@@ -24,8 +23,7 @@ func BenchmarkContext_TransientStore(b *testing.B) {
 
 	ctx := testutil.DefaultContext(key, types.NewTransientStoreKey("transient_"+b.Name()))
 
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
 		_ = ctx.TransientStore(key)
 	}
 }
@@ -35,8 +33,7 @@ func BenchmarkContext_CacheContext(b *testing.B) {
 
 	ctx := testutil.DefaultContext(key, types.NewTransientStoreKey("transient_"+b.Name()))
 
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
 		_, _ = ctx.CacheContext()
 	}
 }